internal/lsp: remove handling for circular imports
Also, separate type-checking logic into its own file.
go/packages returns import cycle errors anyway, so we just return them instead.

Reviewed-on: https://go-review.googlesource.com/c/tools/+/165021
