--- internal/lsp/testdata/self/self.go.in
@@ -1,13 +1,0 @@
-// +build go1.11
-
-package self
-
-import (
-	"golang.org/x/tools/internal/lsp/self" //@diag("\"golang.org/x/tools/internal/lsp/self\"", "LSP", "could not import golang.org/x/tools/internal/lsp/self (import cycle: [golang.org/x/tools/internal/lsp/self])")
-)
-
-func Hello() {}
-
-func _() {
-	self.Hello()
-}
